// ui/src/generated/api/model/dtoTemplateResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export interface DtoTemplateResponse {
  description: string;
  kind: string;
  language: string;
  name: string;
  sourceCode: string;
  title: string;
}